feat: unique default device name and send to CMS on registration
- Add shared/identity.py with device serial helpers (consolidates duplicated code)
- Default device_name is now 'agora-node-XXXX' where XXXX = last 4 of serial
- Send device_name in the CMS register WebSocket message
- Config loads device_name from persist file (set via captive portal)
- Falls back to unique auto-generated name if not explicitly set
