+++ docs/openapi.yaml
@@ -6,7 +6,7 @@ info:
     playback control, CMS integration, and status monitoring. A separate player service
     handles GStreamer pipelines, and an optional CMS client connects to Agora CMS for
     centralized scheduling and fleet management.
-  version: 0.9.0
+  version: 0.9.4
   license:
     name: MIT
